architectures/event-driven/gcp/examples/publisher/requirements.txt
# Pub/Sub Publisher Dependencies
google-cloud-pubsub>=2.21.0
